Move Twitch components into separate folder
