+++ src/renderer/components/ProcessSection.tsx
