+++ src/renderer/components/twitch/index.ts
@@ -1,0 +1,2 @@
+export * from "./TwitchConnectButton";
+export * from "./TwitchUserStatus";
